+++ CHANGELOG.md
@@ -7,7 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2023-04-09
 
+### Added
+
+- Project landing page
+- Project introduction
+- Project getting started guide
+- Project logo
 
 ## [0.1] - 2023-04-07